requirements.txt
asgiref==3.5.0
beautifulsoup4==4.11.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
Django==4.0.4
django-cors-headers==3.11.0
django-dump-load-utf8==0.0.4
django-rest-framework==0.1.0
djangorestframework==3.13.1
filelock==3.6.0
idna==3.3
joblib==1.1.0
lxml==4.8.0
nltk==3.7
numpy==1.22.3
pandas==1.4.2
psycopg2==2.9.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
regex==2022.3.15
requests==2.27.1
six==1.16.0
snscrape==0.4.3.20220107.dev45+ged3ea94
soupsieve==2.3.2.post1
sqlparse==0.4.2
tqdm==4.64.0
tzdata==2022.1
urllib3==1.26.9
